Quick fix for production upload breakage issues.
Upload API requests which time out after 30 seconds were reported
to the user as errors. Increasing the timeout to 90 seconds should
help with the majority of cases we're seeing.

The underlying bug is described at
https://bugzilla.wikimedia.org/show_bug.cgi?id=34785
but is not fully resolved with this as we really need to stop
these requests from exceeding many seconds even for small files.

Change-Id: Ifddb4fa2bfcd7d2959e6014ba33b80d831415711